requirements.txt
dacite
google-auth
google-cloud-run 
google-cloud-secret-manager
python-dotenv 
requests
ujson
